% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_context_segment.R
\name{add_context_segment}
\alias{add_context_segment}
\title{Add Context Segment}
\usage{
add_context_segment(segment, contexts, merge = TRUE, verbose = FALSE)
}
\arguments{
\item{segment}{sequence of two contexts defining a segment to add to 'contexts',
argument in binary long format}

\item{contexts}{a sequence of any number of contexts to which a 'segment'
of contexts will be added via merging with existing contexts or replacing existing
contexts.}

\item{merge}{logical, whether to merge the segment with the
existing sequence or replace existing overlapping contexts, default=FALSE}

\item{verbose}{logical, whether to print out progress messages}
}
\value{
Returns a drop-in context sequence for the passed segment with
appropriate endpoint and middle contexts.
}
\description{
Based on a two-context sequence this function computes the
endpoints and middle points of a corresponding context sequence
to be added to an existing sequence of contexts with the option of replacing or merging
with existing contexts
}
\details{
Contexts are specified in "binary long" format. 'segments' and 'contexts'
must have the same basis of groups and states

Typically the 2nd context in the segment is empty (no active
states) to define a period in which a set of states are active and then
become inactive.  However, this function will still process
segments that have active states at the second context
}
\author{
Neil Klepeis
}
