% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_context_segments.R
\name{add_context_segments}
\alias{add_context_segments}
\title{Add Context Segments}
\usage{
add_context_segments(x, y, replace = TRUE)
}
\arguments{
\item{x}{a sequence of y sequences in binary wide format}

\item{y}{a sequence of y to add to the 'x' sequence
y.}

\item{replace}{logical, whether to merge the x with the
existing sequence or replace existing overlapping y, default=FALSE}
}
\value{
Returns a new sequence that results from adding y to x.
}
\description{
Add one sequence of y to another with the option of
replacing or merging with the existing context sequence
(TBD:  Have option of full left or right join of y)
}
\details{
y are specified in "binary wide" format.  The groups and states
in the returned sequenc are superset of all the groups and states in
x and y.
}
\author{
Neil Klepeis
}
