% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_times_to_contexts.R
\name{add_times_to_contexts}
\alias{add_times_to_contexts}
\title{Add Times to a Context Sequence}
\usage{
add_times_to_contexts(newtimes, contexts)
}
\arguments{
\item{newtimes}{new times to add to the context sequence}

\item{contexts}{a sequence of any number of contexts}
}
\value{
Returns a new context sequence with new duplicate contexts
occurring at specified times
}
\description{
Inserts new contexts into a context sequence at
specified times, duplicating the immediately preceding context values.
Oppposite action of collapsing contexts.
}
\details{
'newtimes' can contain times that are already present and
are silently ignored
}
\author{
Neil Klepeis
}
