% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_stream.R
\name{as_stream}
\alias{as_stream}
\title{Coerce data frame or tibble to a 'stream' data object}
\usage{
as_stream(x)
}
\arguments{
\item{data}{a dataframe or tibble containing Time, Value, and, optionally
other index variables to convert to a 'stream' data type}
}
\description{
Coerce data frame or tibble to a 'stream' data object
}
\details{
Returns a 'stream' data object, a tibble with specific cols types, or
an error if the passed object cannot be coerced to 'stream' data type
}
\author{
Neil Klepeis
}
