% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary2timeline.R
\name{binary2timeline}
\alias{binary2timeline}
\title{Convert an environmental pathway in binary row (wide) format to timeline
format}
\usage{
binary2timeline(
  data,
  time.var = "Time",
  units = "mins",
  ystart = 0,
  yspacing = 0.25,
  yheight = 1,
  sep = ":"
)
}
\arguments{
\item{data}{a date frame containing an environmental pathway (binary wide format)}

\item{time.var}{the expected name for the time column in
'data' (defaults to "time")}

\item{units}{time units to use when encoding elapsed time.}

\item{ystart}{the y coordinate for the origin}

\item{yspacing}{the spacing between state lanes}

\item{yheight}{the height of timeline lanes (defaults to 1)}

\item{state.names}{the state names for an environment we expect to be
present in 'data'}
}
\description{
Convert an environmental pathway in binary row (wide) format to timeline
format
}
\details{
This function converts a data frame containing
an enviroment flow in terms of rows of binary states ("Binary Wide" format) and
a time variable to a timeline-formatted dataframe with xstart,
xend, ystart, and yend for easy plotting.  Heights for each
state are set to unit = 1
}
\author{
Neil Klepeis
}
