% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\title{Get index that collapses a vector to nonrepeated elements}
\usage{
collapse(x)
}
\arguments{
\item{x}{a vector}
}
\description{
Get index that collapses a vector to nonrepeated elements
}
\details{
Returns an index that can be used to collapse the
}
\author{
Neil Klepeis
}
