% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_contexts.R
\name{collapse_contexts}
\alias{collapse_contexts}
\title{Collapse contexts with duplicate states}
\usage{
collapse_contexts(...)
}
\arguments{
\item{...}{one or more sequences of contexts to quantize}
}
\value{
Returns a list of all passed context sequences with
contexts with duplicated states removed
}
\description{
Collapse context sequence so there are no
duplicated contexts in time
}
\seealso{
distinct
}
\author{
Neil Klepeis
}
