% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_contexts.R
\name{combine_contexts}
\alias{combine_contexts}
\title{Combine two or more context sequences}
\usage{
combine_contexts(...)
}
\arguments{
\item{...}{context sequence data frames in any format}
}
\value{
a data frame containing a sequence consisting of merged
context sequences
}
\description{
Combine two or more context sequences into a
single context sequence
}
\details{
Context sequences can be in Binary Long, Binary Wide, Active State, or
Grouped State format.

The combined sequences time range covers the times of all passed sequences.

If times are prior to the range of a given sequences, NA's are inserted for
for states in that context.

If there are no states in common, then the new context sequence contains
all the information in the passed sequences.

If there are groups of states in common among the context sequences to be combined,
then the contexts passed on the right overwrite the state values for
contexts on the left, i.e., the contexts passed later take precedence over
the contexts passed earlier.
}
\examples{

library(tidyverse)
}
\author{
Neil Klepeis
}
