% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_aggregate_time_spent.R
\name{context_aggregate_time_spent}
\alias{context_aggregate_time_spent}
\title{Convert contexts in binary format to time spent in
regular intervals}
\usage{
context_aggregate_time_spent(data, by = "1 hour")
}
\arguments{
\item{data}{a date frame containing contexts in binary long format}
}
\description{
Convert contexts in binary format to time spent in
regular intervals
}
\details{
Strategy

1.  Insert context time points at regular intervals
2.  Iterate over each interval computing total time in each group/state
}
\author{
Neil Klepeis
}
