% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contexts_from_stream_episodes.R
\name{contexts_from_stream_episodes}
\alias{contexts_from_stream_episodes}
\title{Create Contexts from Stream Episodes}
\usage{
contexts_from_stream_episodes(streams, threshold, states)
}
\arguments{
\item{streams}{a tibble of sensor data streams in standard
Time, Response, Value long format}

\item{threshold}{the response value to use as a threshold when
identifying peaks and the extent of each episode}

\item{states}{names of states to assign to the}
}
\value{
a tibble defining contexts
}
\description{
This function creates contexts by
analyzing streams for episodes
}
\details{
A stream episode is defined as a segment of time
in which data stream values continuously occur above a specified
threshold. If multiple responses are present in the passed
value only the first one is used to identify episodes
}
\author{
Neil Klepeis
}
