% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_from_binary_to_groupedstate_contexts.R
\name{convert_from_binary_to_groupedstate_contexts}
\alias{convert_from_binary_to_groupedstate_contexts}
\title{Convert Contexts from Binary to Grouped State Format}
\usage{
convert_from_binary_to_groupedstate_contexts(
  contexts,
  index.vars = NULL,
  sep = ":",
  state_sep = ";"
)
}
\arguments{
\item{contexts}{a data frame containing a sequence of contexts in WIDE binary format}

\item{index.vars}{Index variables to use for grouping (in addition to "Time")}

\item{sep}{delimiter for Groups and States in the WIDE column names}

\item{state_sep}{delimiter for multiple states concurrent for a given group}
}
\value{
A tibble containing a sequence of states in grouped-state format
}
\description{
This function converts a context sequence in WIDE binary format
to grouped-state context format
}
\details{
If states within groups are mutually exclusive, concurrent states
are separated by the 'state_dep' delimiter in the returned tibble.  Having
mutually-exclusive states in each group produces a cleaner context space.

The 'contexts' tibble must have a Time column and multiple columns
named using Group:State pairs, containing 0 or 1 values indicating the occurrence of
a given state at each time.
}
\author{
Neil Klepeis
}
