% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_from_groupedstates.R
\name{convert_from_groupedstates}
\alias{convert_from_groupedstates}
\title{Convert Contexts from Groups}
\usage{
convert_from_groupedstates(contexts, allStates, auto.states = TRUE)
}
\arguments{
\item{contexts}{tibble containing contexts in active-state format}

\item{allStates}{list containing grouped states.
If missing and \code{auto.states = TRUE},
then the function parses columns to assign \code{allStates}
Any states in the data not matched in \code{allStates} will be omitted from the output}

\item{auto.states}{logical, automatically parse columns for grouped state list, defaults to TRUE}
}
\value{
a tibble containing contexts in binary long format
}
\description{
This function converts a data frame of contexts in "grouped state"
format") with columns containing state groups to binary long format
}
\details{
The "grouped state" format is a special kind of "active state" format
in which columns contain occurrence of mutually-exclusive states within
each group. This was essentially the old "tact" format that was designed
with Location and NHAPS activity in mind, which were constrained to be
groups of mutually exclusive states (only a single activity or location
could occurr at a given time). This specification has columns indicating
which states are active for each time (row.)\tabular{lccc}{
   Time \tab Household \tab Entertainment \tab Social \cr
   ________________ \tab _____________ \tab _______________ \tab _________ \cr
   4/2/02 10:12 am \tab Cooking \tab TV \tab Talking \cr
   4/2/02 10:14 am \tab Cooking \tab  \tab Talking \cr
}
}
\author{
Neil Klepeis
}
