% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviation_from_mean.R
\name{deviation_from_mean}
\alias{deviation_from_mean}
\title{Compute deviation from mean for a set of sensors}
\usage{
deviation_from_mean(data, use.percent = TRUE)
}
\arguments{
\item{data}{a data frame in WIDE format with rows containing sensor data
at specific times with each column containing data for an individual sensor}

\item{use.percent}{whether to compute deviations as "relative" (value) or
"percent" values (does percents by default)
central value statistic}
}
\value{
a data frame containing raw deviations from the mean and computed
statistics for deviations from the mean (see Details below)
}
\description{
Compute deviation from the mean for concurrent
sensor data streams with one column per sensor
}
\details{
This function is intended for use in evaluating the agreement between
sensors of the same type by computing deviations of sensor readings from the
mean reading

If a "Time" column is not present in the data frame, then one is added.

All variables not named "Time" are assumed to be streams of sensor data
for the same type of sensor.

Rows with any missing values are dropped.

At each time (row), this function computes:
\itemize{
\item The overall row mean
\item The relative deviation from the mean for each sensor (using same names as input columns)
\item The minimum absolute deviation from the mean across all sensors
\item The maximum absolute deviation from the mean across all sensors
\item The minimum relative deviation from the mean across all sensors
\item The maximum relative deviation from the mean across all sensors
\item The 25th percentile of relative deviations from the mean
\item The 75th percentile of relative deviations from the mean
}
}
\author{
Neil Klepeis
}
