% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviation_from_reference.R
\name{deviation_from_reference}
\alias{deviation_from_reference}
\title{Compute deviation of test sensors from a reference sensor}
\usage{
deviation_from_reference(data, ref = names(data)[1], use.percent = FALSE)
}
\arguments{
\item{data}{a data frame in WIDE format with rows containing sensor data
at specific times with each column containing data for an individual sensor}

\item{ref}{name or number of the column containing the reference data,
defaults to the first column in the data frame}

\item{use.percent}{whether to compute deviations as "relative" (value) or
"percent" values (does percents by default)}
}
\value{
a data frame containing raw deviations from the reference and computed
statistics for deviations from the mean (see Details below)
}
\description{
Compute deviation from a reference for concurrent
sensor data streams with one column per sensor
}
\details{
This function is intended for use in evaluating the agreement between
a number of 'test' sensors of the same type and a 'reference' sensor by computing
deviations of 'test' sensor readings from the 'reference' reading at different
times

If a "Time" column is not present in the data frame, then one is added.

All variables not named "Time" are assumed to be streams of sensor data
with 'ref' containing the name or number of the column containing the reference
data stream

This functions works the same as tte 'deviation_from_mean' function except the
'ref' column takes the place of the mean.

Rows with any missing values are dropped.

At each time (row), this function computes:
\itemize{
\item The overall row mean of 'test' sensors
\item The relative deviation from the reference for each sensor (using same names as input columns)
\item The minimum absolute deviation from the reference across all sensors
\item The maximum absolute deviation from the reference across all sensors
\item The minimum relative deviation from the reference across all sensors
\item The maximum relative deviation from the reference across all sensors
\item The 25th percentile of relative deviations from the reference
\item The 75th percentile of relative deviations from the reference
}
}
\author{
Neil Klepeis
}
