% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multiple_ggplot.R
\docType{data}
\name{df}
\alias{df}
\title{Create some data to play with. Two time series with the same timestamp.}
\format{
An object of class \code{data.frame} with 8761 rows and 3 columns.

An object of class \code{data.frame} with 8761 rows and 3 columns.
}
\usage{
df

df
}
\description{
Create some data to play with. Two time series with the same timestamp.

Create some data to play with. Two time series with the same timestamp.
}
\keyword{datasets}
