% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explorePeaks.R
\name{explorePeaks}
\alias{explorePeaks}
\title{Shiny app to interactively identify peaks in a time series}
\usage{
explorePeaks(streams, maxHeight = 1000, maxDist = 500, maxThreshold = 100)
}
\arguments{
\item{stream}{a tibble containing data streams with
Time, Response, Value columns}
}
\description{
A shiny app to visualize identified peaks in a time series
}
\author{
Neil Klepeis
}
