% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_grouped_states.R
\name{extract_groupstates}
\alias{extract_groupstates}
\title{Extract list of grouped states from a long context object}
\usage{
extract_groupstates(contexts, groupstates = TRUE, sep = ":")
}
\arguments{
\item{contexts}{tibble containing contexts in long format}

\item{groupstates}{logical, whether to return states with embedded groups names or not}
}
\value{
a named list containing states belonging to each group
}
\description{
This function extracts a list with named elements
for each group containing component states for each group
}
\author{
Neil Klepeis
}
