% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_context_episodes.R
\name{find_context_episodes}
\alias{find_context_episodes}
\title{Find Context Episodes}
\usage{
find_context_episodes(contexts)
}
\arguments{
\item{contexts}{contexts in active-state format}
}
\value{
a vector of breaks identifying context episodes
}
\description{
This function identifies episodes from an
active-state context specification
}
\details{
A context episode is defined as a segment of time
in which some combination of states are active in between periods
of no state activity
}
\author{
Neil Klepeis
}
