% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_stream_episodes.R
\name{find_stream_episodes}
\alias{find_stream_episodes}
\title{Find Stream Episodes}
\usage{
find_stream_episodes(
  streams,
  merge = TRUE,
  nups = 1,
  ndowns = 1,
  zero = "0",
  minpeakheight = 0,
  minpeakdistance = 1,
  threshold = 0,
  merge.threshold = Inf,
  sortstr = TRUE,
  by = "30 seconds",
  plot = TRUE,
  returnPlot = FALSE,
  fill = rgb(0.4, 0, 0, 0.15)
)
}
\arguments{
\item{streams}{a tibble of sensor data streams in standard long format
with Time, Response, and Value variables}

\item{nups}{findpeaks argument, min number of increasing steps before the peak, defaults to 1}

\item{ndowns}{findpeaks argument, min number of decreasing steps before the peak, defaults to 1}

\item{minpeakheight}{min height to be identified as a peak, if 0, then auto-assigned as median value}

\item{minpeakdistance}{findpeaks argument, min distance in indices peaks have to be to be counted, defaults to 100}

\item{threshold}{min value identified peaks must be above their neighboring peaks, defaults to 0}

\item{merge.threshold}{threshold value above which all peaks will be
merged into a single peak, defaults to Inf.}

\item{sortstr}{findpeaks argument, returned peaks sorted in decreasing order?, defaults to TRUE}

\item{by}{the time interval to aggregrate the stream data for
some pre-smoothing before the loess smoothing, defaults to "30 secs"}

\item{plot}{logical, whether to make a plot of the found episodes}

\item{returnPlot, }{logical, whether to return the plot or the data}

\item{zerofindpeaks}{argument, how to interpret succeeding steps of the same value, defaults to 0}
}
\value{
a dataframe defining the identified stream episodes or, if merge.threshold
is not Inf or -Inf, as list with raw peaks data frame component and a mergedPeaks data frame
component
}
\description{
This function identifies episodes in a
stream of sensor data using the pracma findpeaks function
and loess pre-smoothing of the data
}
\details{
A stream episode is defined as a segment of time
in which data stream values rise from baseline to a peak and
fall back to baseline. If multiple responses are present in the passed
value only the first one is used to identify episodes.
The baseline is taken as the median of smoothed values.

Merging of peaks is supported by specifying a 'merge.threshold' that is
not Inf or -Inf.  Groups of peaks that continuously stay above the threshold
are merged into a single peak.   This feature is useful to group peaks that seem to
"belong together" in that they are part of an episode where the
levels stay high but may dip up and down -- perhaps due to multiple sources or
an intermittent or unstable source.
}
\author{
David Torres-Mendoza and Neil Klepeis
}
