% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_stream_episodes_thresholds.R
\name{find_stream_episodes_thresholds}
\alias{find_stream_episodes_thresholds}
\title{Find Stream Episodes Using Simple Thresholds}
\usage{
find_stream_episodes_thresholds(
  streams,
  lower,
  upper,
  context = TRUE,
  states = list(Episodes = "Episode")
)
}
\arguments{
\item{streams}{a tibble of sensor data streams in standard
Time, Response, Value long format}

\item{lower}{the response value to use as a lower bound when
identifying peaks and the extent of each episode}

\item{upper}{the response value to use as an upper bound when
identifying peaks and the extent of each episode}

\item{context}{if TRUE return in active-state context format
other return a tibble with start/end times for each episode}

\item{states}{a list of with names consisting of groups and
elements consisting of state names within each group}
}
\value{
a vector of breaks identifying stream episodes
}
\description{
This function identifies episodes in a
stream of sensor data
}
\details{
A stream episode is defined as a segment of time
in which data stream values contiuously occur above a specified
threshold. If multiple responses are present in the passed
value only the first one is used to identify episodes
}
\author{
Neil Klepeis
}
