% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_stream_episodes.R
\name{fine_stream_episodes}
\alias{fine_stream_episodes}
\alias{find_stream_episodes}
\title{Find Stream Episodes}
\usage{
find_stream_episodes(
  streams,
  nups = 1,
  ndowns = 1,
  zero = "0",
  minpeakdistance = 1,
  thresh = 0,
  npeaks = 10,
  sortstr = TRUE,
  by = "30 seconds",
  plot = TRUE
)
}
\arguments{
\item{streams}{a tibble of sensor data streams in standard long format
with Time, Response, and Value variables}

\item{nups}{findpeaks argument, min number of increasing steps before the peak, defaults to 1}

\item{ndowns}{findpeaks argument, min number of descreasing steps before the peak, defaults to 1}

\item{minpeakdistance}{findpeaks argument, min distance in indices peaks have to be to be counted, defaults to 100}

\item{thresh}{findpeaks argument, min value to be counted as a peak, defaults to 0}

\item{npeaks}{findpeaks argument, number of peaks to return, defaults to 10}

\item{sortstr}{findpeaks argument, returned peaks sorted in decreasing order?, defaults to TRUE}

\item{by}{the time interval to aggregrate the stream data for
some pre-smoothing before the loess smoothing, defaults to "30 secs"}

\item{plot}{logical, whether to make a plot of the found episodes}

\item{zerofindpeaks}{argument, how to intrepret succeeding steps of the same value, defaults to 0}
}
\value{
a dataframe defining the identified stream episodes
}
\description{
This function identifies episodes in a
stream of sensor data using the pracma findpeaks function
and loess pre-smoothing of the data
}
\details{
A stream episode is defined as a segment of time
in which data stream values rise from baseline to a peak and
fall back to baseline. If multiple responses are present in the passed
value only the first one is used to identify episodes.
The baseline is taken as the median of smoothed values.
}
\author{
David Torres-Mendoza and Neil Klepeis
}
