% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_log_probability_bins.R
\name{geom_log_probabilility_bins}
\alias{geom_log_probabilility_bins}
\alias{geom_log_probability_bins}
\title{Add binned data to log probability plot}
\usage{
geom_log_probability_bins(h, l, size = 2.3, color = "red", ...)
}
\arguments{
\item{...}{arguments to geom_qq function}
}
\description{
Add binned data to an exisiting log-probability plot
}
\details{

}
\author{
Neil Klepeis
}
