% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_log_probability_data.R
\name{geom_log_probabilility_data}
\alias{geom_log_probabilility_data}
\alias{geom_log_probability_data}
\title{Add data to log probability plot}
\usage{
geom_log_probability_data(...)
}
\arguments{
\item{...}{arguments to geom_qq function}
}
\description{
Add raw data to an exisiting log-probability plot
}
\author{
Neil Klepeis
}
