% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_log_probability_lnorm.R
\name{geom_log_probabilility_lnorm}
\alias{geom_log_probabilility_lnorm}
\alias{geom_log_probability_lnorm}
\title{Add log-normal distribution to log probability plot}
\usage{
geom_log_probability_lnorm(gm = 1, gsd = exp(1), ...)
}
\arguments{
\item{...}{arguments to geom_line function}
}
\description{
Add log-normal distribution to an exisiting log-probability plot
}
\details{

}
\author{
Neil Klepeis
}
