% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_log_probability_abline.R
\name{geom_log_probability_abline}
\alias{geom_log_probability_abline}
\title{Add log-normal distribution to log probability plot}
\usage{
geom_log_probability_abline(
  mapping = NULL,
  data = NULL,
  ...,
  gm,
  gsd,
  na.rm = FALSE,
  show.legend = NA
)
}
\arguments{
\item{gm}{geometric mean for the line}

\item{gsd}{geometric standard deviation for the line}

\item{show.legend}{}
}
\description{
Add log-normal distribution to an exisiting log-probability plot
}
\details{
This geom is an extension of the \code{geom_abline} that takes gm and gsd
parameters and computes appropriate slope and intercept to plot the
line on a log-probability axis, representing a log-normal probability
distribution
}
\author{
Neil Klepeis
}
