% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_allStates.R
\name{get_allStates}
\alias{get_allStates}
\title{Get all existing and selected context states}
\usage{
get_allStates(contextsLong, contextSets)
}
\arguments{
\item{contextsLong}{a data frame containing long-format contexts}

\item{contextSets}{a list containing states from selected context sets}
}
\description{
Merge states collected from existing long-format contexts
and a list of selected states
}
\author{
Neil Klepeis
}
