% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist2density.R
\name{hist2raw}
\alias{hist2raw}
\alias{hist2density}
\title{Simulate raw values from a counts and limits defining a histogram}
\usage{
hist2density(counts, breaks, simulate = TRUE)
}
\arguments{
\item{counts}{number of observation in each bin as defined by breaks}

\item{breaks}{histogram breaks (class interval limits)}
}
\value{
a tibble containing raw data simulated from a histogram
}
\description{
Simulate raw values from a counts and limits defining a histogram
}
\details{
Raw data are simulated in each histogram bin assuming
a uniform distribution withing the range of values defined by the
limits of each bin.

The output of this function can be used to compute a kernel density estimation
to plot a smoothed histogram
}
\author{
Neil Klepeis
}
