% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_join_by_time.R
\name{left_join_by_time}
\alias{left_join_by_time}
\title{Join a data frame to another according to row matches by time}
\usage{
left_join_by_time(x, y, by = NULL)
}
\arguments{
\item{x}{a data frame with a Time variable}

\item{y}{a data frame with a Time variable with columns to "join" or
merge with columns of 'y' matching by time interval}

\item{by}{extra variables to merge by (in addition to "Time"). See \link{\code{left_join}}}
}
\value{
a data frame with all columns of 'x' and columns from 'y' that
match time intervals in 'x'
}
\description{
Take two data frames with time varying variables and
merge the second into the first based on the first's times falling
into the row-defined time intervals of the second
}
\details{
First find the times of rows in in 'x' that are encompassed by rows
rows in 'y' (reassigning 'y' times to those of the 'x' time match), then
get distinct rows and do a normal tidyverse 'left_join'
to merge the 'y' rows into the matching 'x' rows
}
\examples{

library(tidyverse)
d <- tibble(Time=c(1,1,1,2,2,2,5,5,5,6,6),A=c(5,3,6,4,3,3,7,8,5,8,3))
k <- tibble(Time=c(1.5,2.2,3,4.1,4.1,5,5,5.3),B=c(10,20,40,50,60,70,80,90))
left_join_by_time(d,k)
}
\author{
Neil Klepeis
}
