% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_files.R
\name{list_data_files}
\alias{list_data_files}
\title{Get a list of data files for processing}
\usage{
list_data_files(
  prefixes = c("CNR", "IPS", "SidePak", "DustTrak"),
  extension = "csv"
)
}
\arguments{
\item{prefixes}{defaults ot "CNR", "IPS", "PurpleAir", "SidePak", "DustTrak"}

\item{extension}{defaults to "csv"}
}
\value{
list of full data file names matching prefix and extension
}
\description{
Look in current directory for files with specified prefixes
and extension
}
\details{
Looks recursively in sub-directories and returns full names
with the file path relative to current directory
}
\author{
Neil Klepeis
}
