% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_tics.R
\name{log_tics}
\alias{log_tics}
\title{Compute tics for a log10 axis}
\usage{
log_tics(x, exact10 = TRUE)
}
\arguments{
\item{x}{a vector of data or data range for which tics will be created}

\item{exact10, }{logical, whether to compute exact base-10 tics}
}
\value{
a list containing the input data range (from x), the calculated tic range,
and the calculated major and minor tic values.
}
\description{
Find nice log base-10 axis tics.  Values in x less than 0 are removed.
Can set axis tics to exact multiples of ten, e.g., 0.1, 1, 10, if \code{exact10},
or nearest 10^i increment, e.g., 0.2 or 0.3 (for i = -1) otherwise
}
\author{
Neil Klepeis
}
