% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_time_segments.R
\name{match.time.segments}
\alias{match.time.segments}
\alias{match_time_segments}
\title{Calculate factor indicating time segements}}
\usage{
match_time_segments(
  times,
  segment.limits,
  labels,
  format,
  add.integers = FALSE,
  sep = "-",
  collapse = TRUE,
  debug = FALSE
)
}
\arguments{
\item{times}{a vector of POSIXct times, or character representation of times}

\item{segment.limits}{a vector of POSIXct time limits for the segments, or a character representation of times}

\item{labels}{optional, labels corresponding to the time segments, must be 1 element smaller than the time segment vector.  Given integer values by default.}

\item{format}{optional, giving the format of the character representation of times and time segments}
}
\value{
Returns a factor containing levels given in `labels' showing the specified time segments
}
\description{
Take a vector of times and limits for time segments and create a factor that indicates the time segments
}
\keyword{manip}
