% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_by_time.R
\name{merge_by_time}
\alias{merge_by_time}
\title{Merge Data by Time Interval}
\usage{
merge_by_time(data, keys)
}
\arguments{
\item{data}{time-varying data, with Time variable and any number of other key or response variables}

\item{key}{time-varying data with key variables to merge into \code{data}}
}
\value{
a tibble containing data from \code{keys} merged into \code{data}
}
\description{
This function provides a general way to merge data frames
with time-varying variables by matching individual times in \code{keys} with
time intervals in \code{data}, i.e., there is a match when an individual time
falls within a given time interval.
}
\details{
This function expects both \code{data} and \code{key} to have a POSIXt column
named \code{Time}.  All other columns are retained in the merged result.

The time in each row in the \code{keys} data frame is matched to
one or more \emph{time intervals} in the \code{data} data frame.  Matching is defined by
the individual times in \code{keys} falling within time intervals defined by
successful times in \code{data}.

Multiple rows in \code{keys} may match one or more rows in \code{data} so the
results may have duplicates of rows in either input data frames, i.e., a full_join
type of merging in the tidyverse.

IMPORTANT:  The underlying assumption is that the times in \code{data} correspond
to instantaneous time with associated values and the time in \code{keys} correspond
to flags that indicate the activity of other variables values at a momment in
time, with activity changing at eqch subsequent time.   If \code{data} times are
too far apart (i.e., wider than intervals between successive times in \code{keys}), then
no data in \code{data} will be flagged with key (activity) data.

TODO:  Do a left or right join so that duplicates of keys or data are omitted.
}
\examples{

library(tidyverse)
d <- tibble(Time=c(1,1,1,2,2,2,5,5,5,6,6),A=c(5,3,6,4,3,3,7,8,5,8,3))
k <- tibble(Time=c(1.5,2.2,3,4.1,4.1,5,5,5.3),B=c(10,20,40,50,60,70,80,90))
merge_by_time(d,k)
}
\author{
Neil Klepeis
}
