% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_streams_and_contexts.R
\name{merge_streams_and_contexts}
\alias{merge_streams_and_contexts}
\title{Merge long-format sensor data streams and contexts}
\usage{
merge_streams_and_contexts(streams, contexts, include.binary = FALSE)
}
\arguments{
\item{streams}{Long-format sensor data frame for multiple streams
containing columsn for Time, Response and Value, and optionally other index columns.}

\item{contexts}{Long-format data frame containing binary contextual data consisting of
Time, Group, State, and Value columns.}

\item{include.binary}{logical, whether to include columns for each context state
containing binary values of each state (active = 1; inactive = 0).}
}
\value{
A data frame consisting of the steams data frame with added
columns containing active-state format context data time-matched to each
each data streams row.  Unmatched times in the streams are dropped.
}
\description{
Merge long-format sensor data streams and contexts
}
\details{
This function merges discrete context data into data streams
(continuous time series) data by time-matching contexts to data streams.

Two index columns are added, one with context states separated by "|" and
one with a new line character "\n" used to delineate context states

NOTE: Times not having any active states in the 'streams' data are
assigned an empty context:  "" for active-state format and all 0's for binary format.
These are converted to NA's, which are dropped from the output.

IMPORTANT:  If the streams time resolution is equal or smaller
than, and an even divisor of the context time resolution, then
multiple contexts cannot occur between stream intervals
(i.e., causing context information to be lost).
}
\seealso{
The 'left_join_by_time' function seems to be more efficient (and reliable)
and based on the existing tidyverse left_join function.
}
\author{
Neil Klepeis
}
