% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtact.R
\name{mtact}
\alias{mtact}
\title{Create a multi time-activity object}
\usage{
mtact(..., names = NULL)
}
\description{
Create an mtact object
}
\details{
Creates a multiple time-activity object (mtact) from a number
of single time-activity (tact) objects (or coercable lists)
or a list of tact objects (or coercable lists).
Individual tact objects can be named either as \if{html}{\out{<name>}}=list(...), or
by using the `names' argument, which takes precedent.
}
\author{
Neil Klepeis
}
