% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_mtact.R
\name{panel_mtact}
\alias{panel_mtact}
\title{lattice panel function for plotting multiple time activity data groups (contexts)}
\usage{
panel_mtact(
  z,
  codes = NULL,
  codes.colors = NULL,
  alpha.tact = 0.15,
  fill = TRUE,
  pattern = FALSE,
  overlap.bars = FALSE,
  overlap.all = FALSE,
  show.labels = FALSE,
  labels = NULL,
  srt.label = -90,
  adj.label = c(0, 0),
  fontsize = 0.5,
  fontfamily = "HersheySans",
  fontface = "bold",
  font.col = "black",
  col.border = 0,
  lwd.border = 1,
  box = FALSE,
  col.box = 1,
  lwd.box = 1,
  pattern.type = "segment",
  pattern.angle = 45,
  pattern.width = 3,
  pattern.col = "black",
  pattern.pch = 16,
  pattern.cex = 0.5,
  margin.bottom = 0,
  margin.top = 0
)
}
\description{
This function provides a function to plot multiple time activity context
timelines on a lattice plot
}
\author{
Neil Klepeis
}
