% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_superpose_tact.R
\name{panel_superpose_tact}
\alias{panel_superpose_tact}
\title{lattice panel function for superimposing time activity (context) data on a lattice xyplot}
\usage{
panel_superpose_tact(
  x,
  y,
  subscripts,
  groups = NULL,
  tact.data,
  tact.var,
  col = NULL,
  col.line = superpose.symbol$col,
  col.symbol = superpose.symbol$col,
  pch = superpose.symbol$pch,
  cex = superpose.symbol$cex,
  cex.key = 0.8,
  columns.key = 2,
  between.key = 1,
  between.columns = 0,
  lty = superpose.line$lty,
  lwd = superpose.line$lwd,
  type = "p",
  group.factor = 1,
  border.area = trellis.par.get("plot.polygon")$border,
  alpha.area = trellis.par.get("plot.polygon")$alpha,
  fill.area = trellis.par.get("plot.polygon")$col,
  origin.area = 0,
  remove.nonincreasing = FALSE,
  show.inside.key = TRUE,
  debug = FALSE,
  abline = NULL,
  ...
)
}
\description{
This function provides a function to define a conditioning panel for
a lattice plot in which continuous sensor data streams are combined with timeline data
on activities and events
}
\author{
Neil Klepeis
}
