% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multiple_ggplot.R
\docType{data}
\name{plot1}
\alias{plot1}
\title{Create the two plots.}
\format{
An object of class \code{gg} (inherits from \code{ggplot}) of length 9.

An object of class \code{gg} (inherits from \code{ggplot}) of length 9.
}
\usage{
plot1

plot1
}
\description{
Create the two plots.

Create the two plots.
}
\keyword{datasets}
