% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barchart_plotly2.R
\name{plot_barchart_plotly2}
\alias{plot_barchart_plotly2}
\title{Plot barcharts summarizing sensor data streams}
\usage{
plot_barchart_plotly2(
  data,
  averaging = NULL,
  FUN = "median",
  title = NULL,
  height = 500,
  legend.inside = FALSE,
  fg = "white",
  bg = "rgba(23,103,124,1)",
  plot.bg = "gray90",
  legend.bg = "rgba(255,255,255,0.75)",
  legend.fg = "black",
  line.width = 1
)
}
\arguments{
\item{data}{Long-format sensor data frame for multiple streams and monitoring sites
containing a 'Time' column, index (grouping) columns, and Response and Value columns}

\item{averaging}{averaging time used to aggregate the data streams}

\item{FUN}{function used to compute a statistic for each stream}

\item{title}{title of the plot}

\item{height}{height of the plot in pixels}

\item{fg}{foreground color for text}

\item{bg}{background color}

\item{plot.bg}{plot background color}

\item{legend.bg}{legend background color}

\item{legend.fg}{legend foreground color}

\item{line.width}{linewidth for plotted series}
}
\value{
a plotly object
}
\description{
Plot barcharts for different statistics of data streams
}
\details{
This function plots a barchart comparing a selected
statistic computed from the sensor data streams
}
\author{
Neil Klepeis
}
