% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_barplot_contextual_streams.R
\name{plot_barplot_contextual_streams}
\alias{plot_barplot_contextual_streams}
\title{Barplot of Contextual Data Streams for Current WorkSpace}
\usage{
plot_barplot_contextual_streams(
  streams,
  contexts,
  bg = "white",
  fg = "black",
  date_angle = 30,
  height = 600
)
}
\arguments{
\item{streams}{data frame in long format containing sensor data streams}

\item{contexts}{data frame containing contextual data in "active-state" format for each
defined workspace}
}
\value{
Returns a list of lists containing plots for current
workspace and combination of contextual states
}
\description{
Create a barplot of medians for data streams broken out by contextual states for
the current data-stream workspace
}
\details{
IN FUTURE:  (TODO)

Plot by each combination of states

Aggregate:
\enumerate{
\item Within workspace, across state groups (exposure from mobile site/recpeotr)
\item Within state grouping, across workspace (exposure from fixed sites)
}
}
\author{
Neil Klepeis
}
