% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combo_context_streams2.R
\name{plot_combo_context_streams2}
\alias{plot_combo_context_streams2}
\title{Combo Plot with Contexts and Streams v2}
\usage{
plot_combo_context_streams2(
  timeline,
  streams,
  xrange = NULL,
  height = NULL,
  showlegend = TRUE,
  showTimelineLabels = FALSE,
  fg = "white",
  bg = "#2a8094",
  legend.bg = bg,
  legend.fg = fg,
  timeline.font.size = 12,
  heights = c(0.7, 0.3),
  collapse = FALSE,
  include.rangeSelector = FALSE,
  include.rangeSlider = FALSE,
  ...
)
}
\arguments{
\item{timeline}{timeline format contexts (group, state, xleft, xright)}

\item{streams}{Long-format data streams (Time, Response, Value)}

\item{xrange}{range for the xaxis, used to specify zoom level}

\item{fg}{foreground color}

\item{bg}{background color}

\item{collapse}{logical, whether to collapse grouped states into a single lane}

\item{displayModeBar}{logical, whether to show the mode bar for the plot}
}
\description{
This function create a plotly combination plot (subplot)
containing both context and stream data
}
\details{
This function is a rewrite of v. 1 to be more convenient, taking
more standard long-format contexts and long-format sensors stream data
and creating a nice combo plot
}
\author{
Neil Klepeis
}
