% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combo_context_streams_ggplot.R
\name{plot_combo_context_streams_ggplot}
\alias{plot_combo_context_streams_ggplot}
\title{Combo Plot with Contexts and Streams Using \code{ggplot}}
\usage{
plot_combo_context_streams_ggplot(
  contexts,
  streams,
  contexts.args = list(),
  streams.args = list(),
  heights = c(0.5, 0.5)
)
}
\arguments{
\item{contexts}{Long-format binary contexts (Time, Group, State, Value)}

\item{streams}{Long-format data streams (Time, Response, Value)}

\item{contexts.args}{list of named arguments to pass to the \code{plot_state_timeline} function}

\item{streams.args}{list of named argments to pass to the \code{plot_streams_ggplot} function}

\item{heights}{2-element vector with elements in \link{0,1} giving the relative heights of the plots}
}
\description{
This function create a ggplot combination (multiple) plot
containing both context and stream data
}
\details{
This function is similar to the plotly version that plots context data
in a separate but datetime-axis-linked below the sensor stream data.
}
\author{
Neil Klepeis
}
