% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combo_context_streams_plotly.R
\name{plot_combo_contexts_streams_plotly}
\alias{plot_combo_contexts_streams_plotly}
\title{Combo Plot with Contexts and Streams v2}
\usage{
plot_combo_contexts_streams_plotly(
  contexts,
  streams,
  fg = "white",
  bg = "#2a8094",
  font.size = 12,
  heights = c(0.7, 0.3),
  collapse = FALSE,
  include.rangeSelector = FALSE,
  include.rangeSlider = FALSE
)
}
\arguments{
\item{contexts}{Long-format binary contexts (Time, Group, State, Value)}

\item{streams}{Long-format data streams (Time, Response, Value)}

\item{fg}{foreground color}

\item{bg}{background color}

\item{collapse}{logical, whether to collapse grouped states into a single lane}

\item{displayModeBar}{logical, whether to show the mode bar for the plot}
}
\description{
This function create a plotly combination plot (subplot)
containing both context and stream data
}
\details{
This function is a rewrite of v. 1 to be more convenient, taking
more standard long-format contexts and long-format sensors stream data
and creating a nice combo plot
}
\author{
Neil Klepeis
}
