% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_combo_episodes_plotly.R
\name{plot_combo_episodes_plotly}
\alias{plot_combo_episodes_plotly}
\title{Create stacked plots of grouped combo timeline/streams data}
\usage{
plot_combo_episodes_plotly(
  timeline,
  streams,
  breaks,
  ncols = 3,
  height = NULL,
  heights = c(0.7, 0.3),
  bg = "white",
  fg = "black",
  collapse = FALSE,
  ...,
  displayModeBar = TRUE
)
}
\arguments{
\item{timeline}{a dataframe containing group, state, xleft, xright}

\item{streams}{Long-format data streams (Time, Response, Value)}

\item{breaks}{a dataframe with start/end columns defining episodes}

\item{ncols}{number of columns for plotting panels of episodes}

\item{bg}{background color}

\item{fg}{foreground color}

\item{displayModeBar}{logical, whether to show the mode bar for the plot}
}
\description{
This function creates a grid of plotly combination
with individual subplots containing subplots of timelines and streams
for each define episode
}
\details{
The subplots are stacked on top of one another
}
\author{
Neil Klepeis
}
