% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contexts_hourly_aggregate.R
\name{plot_contexts_hourly_aggregate}
\alias{plot_contexts_hourly_aggregate}
\title{Plot column chart showing time spent in contexts by hour
aggregated across all days}
\usage{
plot_contexts_hourly_aggregate(
  data,
  x.lab = "Hour",
  y.lab = "Value",
  title = "Sensor Contexts Hourly Means",
  breaks = c(0, 6, 12, 18),
  labels = c("Midnight", "6a", "Noon", "6pm")
)
}
\arguments{
\item{data}{Data frame in long binary context format containing
variables for 'Time','Group', and 'State'}
}
\value{
a ggplot object
}
\description{
Plot column chart showing time spent in contexts by hour
aggregated across all days
}
\details{
This function produces a column chart showing the mean
level of the data stream by hour of the day, aggregated across
all available days.
}
\author{
Neil Klepeis
}
