% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist_density.R
\name{plot_hist_density}
\alias{plot_hist_density}
\title{Plot a density over an empirical histogram}
\usage{
plot_hist_density(
  counts,
  breaks,
  show.hist = TRUE,
  show.density = TRUE,
  show.freqpoly = TRUE,
  fill.freq = TRUE,
  trans = "log10",
  simulate = TRUE
)
}
\arguments{
\item{counts}{number of observation in each bin as defined by breaks}

\item{breaks}{histogram breaks (class interval limits)}

\item{show.hist}{logical, whether to also plot the histogram}
}
\description{
Plot a density over an empirical histogram
}
\details{
This function plots a smoothed histogram using
kernel density estimation.  Optionally the original histogram
is also drawn.

Alternatives
\itemize{
\item Just use a Frequencey Plot (geom_poly) to connect all the histogram points
\item Just overlap the histograms using an alpha channel
\item Use the ggridges (density ridgeline plot) to show how distribution changes over time/size/etc.
\item Use splines... or bezier
}

ggridges is AWESOME!!  Use for showing probability distributions changing
over time or as a function of source, distance, etc.

We can do several things to characterize a distribution:
\itemize{
\item histogram
\item frequency plot
\item density esimate
\item mode, median, mean, sd, gsd
\item quartiles...
}
}
\author{
Neil Klepeis
}
