% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_histograms_plotly2.R
\name{plot_histograms_plotly2}
\alias{plot_histograms_plotly2}
\title{Plot histograms for sensor data streams}
\usage{
plot_histograms_plotly2(
  data,
  logaxis = TRUE,
  averaging = NULL,
  title = NULL,
  height = 500,
  legend.inside = FALSE,
  fg = "white",
  bg = "rgba(23,103,124,1)",
  plot.bg = "gray90",
  legend.bg = "rgba(255,255,255,0.75)",
  legend.fg = "black",
  line.width = 1,
  marker.size = 8
)
}
\arguments{
\item{data}{Long-format sensor data frame for multiple streams and monitoring sites
containing a 'Time' column, index (grouping) columns, and Response and Value columns}

\item{logaxis}{whether to plot histogram(s) with a log X axis}

\item{averaging}{averaging time used to aggregate the data streams}

\item{title}{title of the plot}

\item{height}{height of the plot in pixels}

\item{fg}{foreground color for text}

\item{bg}{background color}

\item{plot.bg}{plot background color}

\item{legend.bg}{legend background color}

\item{legend.fg}{legend foreground color}

\item{line.width}{linewidth for plotted series}
}
\value{
a plotly object
}
\description{
Plot histograms for all sites and their selected responses
}
\details{
This function plots overlapping histograms for all responses
and all sites in the passed data frame
}
\author{
Neil Klepeis
}
