% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_historical_data_plotly.R
\name{plot_historical_data_plotly}
\alias{plot_historical_data_plotly}
\title{Plot historical sensor data using the plotly package}
\usage{
plot_historical_data_plotly(
  data,
  title = NULL,
  showlegend = TRUE,
  shape = "hv",
  source = "source",
  fg = "white",
  bg = "rgba(23,103,124,1)",
  register = TRUE,
  include.rangeSelector = FALSE,
  include.rangeSlider = FALSE
)
}
\arguments{
\item{data}{Data frame containing a 'Time' column and columns
containing the sensor data series}
}
\value{
a plotly object
}
\description{
Plot historical sensor data using the plotly package
}
\author{
Neil Klepeis
}
