% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_historical_data_plotly2.R
\name{plot_historical_data_plotly2}
\alias{plot_historical_data_plotly2}
\title{Plot LONG historical sensor data using the plotly package}
\usage{
plot_historical_data_plotly2(
  data,
  area = TRUE,
  title = NULL,
  showlegend = TRUE,
  legend.inside = TRUE,
  height = NULL,
  xrange = NULL,
  yrange = NULL,
  shape = "hv",
  mode = "lines",
  source = "source",
  fg = "white",
  bg = "rgba(23,103,124,1)",
  plot.bg = "gray90",
  legend.bg = "rgba(255,255,255,0.75)",
  legend.fg = "black",
  legend.font.size = 11,
  axis.font.size = 10,
  line.width = 1,
  marker.size = 6,
  reverse = FALSE,
  register = TRUE,
  include.rangeSelector = FALSE,
  include.rangeSlider = FALSE
)
}
\arguments{
\item{data}{Long-format sensor data frame for multiple streams frame containing a 'Time' column, index columns,
and Response and Value columns}

\item{include.rangeSlider}{}
}
\value{
a plotly object
}
\description{
Plot LONG historical sensor data using the plotly package
}
\details{
concatenates all columns not Time or Value to
a single index variable that is used to plot data series.

See 'plot_historical_subplot_plotly.R' which calls this function
to produce stacked subplots for each non-Response index variable, i.e.,
all index variables besides Response are used to facet the data into
separated plots
}
\author{
Neil Klepeis
}
