% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_historical_dataL.R
\name{plot_historical_dataL}
\alias{plot_historical_dataL}
\title{Plot historical sensor data using the plotly package, L version}
\usage{
plot_historical_dataL(
  data,
  device.name = "DeviceName",
  time.name = "Time",
  value.name = "Value",
  response.name = "Response",
  device.group = "DeviceGroup",
  title = NULL,
  showlegend = TRUE,
  displayModeBar = TRUE,
  shape = "hv",
  area = TRUE,
  line.width = 1,
  source = "source",
  fg = NULL,
  bg.plot = NULL,
  bg.paper = NULL,
  legend.x = 0.5,
  legend.y = 100,
  legend.orientation = "h",
  register = TRUE,
  include.rangeSelector = FALSE,
  include.rangeSlider = FALSE,
  visible.x = TRUE,
  visible.y = TRUE,
  trace.colors = NULL,
  traceorder = "normal"
)
}
\arguments{
\item{data}{Long format data frame containing a 'Time' column variables
containing the sensor data series}
}
\value{
a plotly object
}
\description{
Plot historical sensor data using the plotly package, L version
}
\author{
Neil Klepeis
}
