% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_historical_subplots_plotly.R
\name{plot_historical_subplots_plotly}
\alias{plot_historical_subplots_plotly}
\title{Create stacked plots of grouped data streams}
\usage{
plot_historical_subplots_plotly(
  data,
  shareX = TRUE,
  shareY = FALSE,
  ncols = 2,
  paper.bg = "white",
  ...,
  displayModeBar = TRUE
)
}
\arguments{
\item{data}{a dataframe containing 'Time', 'Response', and 'Value' columns
and at least one additional index column to be used in creating stacked subplots}

\item{displayModeBar}{logical, whether to show the mode bar for the plot}

\item{fg}{foreground color}

\item{bg}{background color}
}
\description{
This function creates a stacked plotly combination plot
with individual subplots for each unique set of index values
}
\details{
The subplots are stacked on top of one another
}
\author{
Neil Klepeis
}
