% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_log_probability_plotly2.R
\name{plot_log_probability_plotly2}
\alias{plot_log_probability_plotly2}
\title{Plot log-probability for sensor data streams}
\usage{
plot_log_probability_plotly2(
  data,
  xtics = c(1e-04, 0.001, 0.05, 0.1, 0.25, 0.5, 0.75, 0.9, 0.95, 0.999, 0.9999),
  averaging = NULL,
  title = NULL,
  height = 500,
  legend.inside = FALSE,
  fg = "white",
  bg = "rgba(23,103,124,1)",
  plot.bg = "gray90",
  legend.bg = "rgba(255,255,255,0.75)",
  legend.fg = "black",
  line.width = 1,
  marker.size = 8
)
}
\arguments{
\item{data}{Long-format sensor data frame for multiple streams and monitoring sites
containing a 'Time' column, index (grouping) columns, and Response and Value columns}

\item{xtics}{the tics to use on the x (probability) axis}

\item{averaging}{averaging time used to aggregate the data streams}

\item{title}{title of the plot}

\item{height}{height of the plot in pixels}

\item{fg}{foreground color for text}

\item{bg}{background color}

\item{plot.bg}{plot background color}

\item{legend.bg}{legend background color}

\item{legend.fg}{legend foreground color}

\item{line.width}{linewidth for plotted series}

\item{marker.size}{symbol size for plotted series}
}
\value{
a plotly object
}
\description{
Plot log-probability for selected responses
}
\details{
This function plots data stream values on log-probability axes
for all responses and all sites in the passed data frame
}
\author{
Neil Klepeis
}
