% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map_leaflet.R
\name{plot_map_leaflet}
\alias{plot_map_leaflet}
\title{Plot sensor geographic trajectory on a interactive map using leaflet}
\usage{
plot_map_leaflet(data)
}
\arguments{
\item{data}{a tibble of Time, Latitude, and Longitude data}
}
\description{
Plot sensor geographic trajectory on a interactive map using leaflet
}
\details{
This function plots an interactive map showing the trajectory
of sensor GPS locations.
}
\author{
Neil E. Klepeis
}
