% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multiple_historical_dataL.R
\name{plot_multiple_historical_dataL}
\alias{plot_multiple_historical_dataL}
\title{Plot faceted historical sensor data using the plotly package, L version}
\usage{
plot_multiple_historical_dataL(
  data,
  by,
  nrows,
  by.format = NULL,
  shareX = FALSE,
  shareY = TRUE,
  margin = 0.0025,
  device.name = "DeviceName",
  time.name = "Time",
  response.name = "Response",
  value.name = "Value",
  device.group = "DeviceGroup",
  fg = NULL,
  bg.plot = NULL,
  bg.paper = NULL,
  visible.x = FALSE,
  visible.y = FALSE,
  trace.colors = NULL,
  area = TRUE,
  line.width = 1,
  traceorder = "normal"
)
}
\arguments{
\item{data}{Long format data frame containing a 'Time' column variables
containing the sensor data series}

\item{by}{variable name to use in splitting data for individual facets,
must be a numeric, e.g., date-time variable that is used to sort the
plots in increasing order}

\item{nrows}{how many rows to use in creating subplots}

\item{by.format}{the datetime format to use in printing by value on each plot}

\item{shareX}{logical, whether to share the x axis between subplots, defaults to TRUE}

\item{shareY}{logical, whether to share the y axis between subplots, defaults to TRUE}

\item{margin}{value of plot margin(s) between 0 and 1 (left, right, top, bottom)}

\item{...}{arguments to pass on to the 'plot_historical_data' function}
}
\value{
a plotly object with subplots
}
\description{
Plot faceted historical sensor data using the plotly package, L version
}
\details{
This function creates plotly subplots (facets) for each unique 'by'
value in the data
}
\author{
Neil Klepeis
}
