% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_state_timeline2.R
\name{plot_state_timeline2}
\alias{plot_state_timeline2}
\title{Plot interactive timeline of state activity using plotly}
\usage{
plot_state_timeline2(
  data,
  auto.y = TRUE,
  labels = TRUE,
  xrange = NULL,
  collapse = FALSE,
  yspace = 0.05,
  border = 0.3,
  legend = FALSE,
  date_breaks = waiver(),
  date_labels = "\%I:\%M:\%S \%p \\n \%a \%m/\%d",
  date_angle = 0,
  popup_date = "\%a \%m/\%d/\%y \%I:\%M:\%S \%p",
  bg = "white",
  fg = "black",
  displayModeBar = TRUE,
  height = 450,
  font.size = 12,
  shape.opacity = 0.4,
  shape.linecolor = "lightgray",
  source = "source",
  config = TRUE
)
}
\arguments{
\item{data}{a dataframe containing named columns for
group, state, xleft, xright, ybottom, ytop. x coordinates
are datetime values and y coordinates are numeric indices indicating
vertical height of state lanes}

\item{auto.y}{(logical) whether to automatically compute lane
y coordinates or use ybottom/ytop arguments.}

\item{xrange}{optional, vector of date/times used to specify the
xrange of the plot. If NULL, the xrange defaults to the range inherent in
the min and max time values in the data: c(min(xleft), max(xright)). This
option is used to visualize the contexts withing a broader time period.}

\item{collapse}{(logical) whether to draw grouped states in a single lane or
draw individual state activity in separate lanes.  Collapsing will result
in overlap of state rectangles unless states in the group are
mutually exclusive}

\item{yspace}{vertical space with value [0,1] to add between lanes}

\item{border}{thickness of border around rectangles}

\item{legend}{whether to draw a legend for the 0/1 states}

\item{date_breaks}{a string giving the datetime breaks for the plot
(e.g., "10 mins", "3 hours")}

\item{date_labels}{the format for the datetime axis labels}

\item{date_angle}{angle to position date axis labels}

\item{popup_date}{format for dates in the popups (tooltips)}

\item{bg}{the background color for the plot}

\item{fg}{the foreground color for the plot (text and axis lines)}

\item{displayModeBar}{If using plotly, whether to display a toolbar
over the plot or not}

\item{height}{a number giving the height of the plotly plot in pixels}

\item{source}{string identification for plot events}

\item{config}{logical, whether to configure the plot plot toolbar (may want
to skip this when configuring subplots)}
}
\description{
Plot interactive timeline of state activity using plotly
}
\author{
Neil Klepeis
}
