% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_state_timeline_ggplot.R
\name{plot_state_timeline_ggplot}
\alias{plot_state_timeline_ggplot}
\title{Plot a static timeline of state activity using ggplot}
\usage{
plot_state_timeline_ggplot(
  data,
  ggplot = TRUE,
  fillstate = "groupstate",
  auto.y = TRUE,
  collapse = FALSE,
  lanes = TRUE,
  yspace = 0.05,
  border = 0.3,
  color.border = "lightgray",
  legend = FALSE,
  lab.x = "Time",
  lab.y = "States",
  title = waiver(),
  date_breaks = waiver(),
  date_labels = "\%I:\%M:\%S \%p \\n \%a \%m/\%d",
  xlimits = NULL,
  date_angle = 0,
  popup_date = "\%a \%m/\%d/\%y \%I:\%M:\%S \%p",
  bg = "white",
  fg = "black",
  displayModeBar = TRUE,
  height = 450,
  source = "source",
  config = TRUE
)
}
\arguments{
\item{data}{a timeline format dataframe containing named columns for
group, state, xleft, xright, ybottom, ytop. x coordinates
are datetime values and y coordinates are numeric indices indicating
vertical height of state lanes}

\item{ggplot}{logical, whether to return a ggplot object (default) or convert
from ggplot to plotly object}

\item{fillstate}{the variable to use in defining the state fill color,
either "group" or "groupstate"}

\item{auto.y}{(logical) whether to automatically compute lane
y coordinates or use ybottom/ytop arguments.}

\item{collapse}{(logical) whether to draw grouped states in a single lane or
draw individual state activity in separate lanes.  Collapsing will result
in overlap of state rectangles unless states in the group are
mutually exclusive}

\item{lanes}{logical, added shaded rectangles to highlight lanes for each group/state}

\item{yspace}{vertical space with value \link{0,1} to add between lanes}

\item{border}{thickness of border around rectangles}

\item{color.border}{color of border lines for rectangles}

\item{legend}{whether to draw a legend for the 0/1 states}

\item{lab.x}{x axis label}

\item{lab.y}{y axis label}

\item{date_breaks}{a string giving the datetime breaks for the plot
(e.g., "10 mins", "3 hours")}

\item{date_labels}{the format for the datetime axis labels}

\item{date_angle}{angle to position date axis labels}

\item{popup_date}{format for dates in the popups (tooltips)}

\item{bg}{the background color for the plot}

\item{fg}{the foreground color for the plot (text and axis lines)}

\item{displayModeBar}{If using plotly, whether to display a toolbar
over the plot or not}

\item{height}{a number giving the height of the plotly plot in pixels}

\item{source}{string identification for plot events}

\item{config}{logical, whether to configure the plot toolbar (may want
to skip this when configuring subplots)}
}
\description{
This function creates a static timeline plot of context states
using the ggplot2 library with the option of converting to an interactive
timeline using the \code{ggplotly} function in the plotly library
}
\details{
This function creates a static state timeline plot using the ggplot2 library.
It has an option to convert from a ggplot object to an interactive plotly object using the
\code{ggplotly} function.
}
\author{
Neil Klepeis
}
