% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_state_timeline_plotly.R
\name{plot_state_timeline_plotly}
\alias{plot_state_timeline_plotly}
\title{Plot an interactive timeline of state activity using plotly}
\usage{
plot_state_timeline_plotly(
  data,
  auto.y = TRUE,
  labels = TRUE,
  xrange = NULL,
  collapse = FALSE,
  yspace = 0.05,
  legend = FALSE,
  date_breaks = waiver(),
  date_labels = "\%I:\%M:\%S \%p \\n \%a \%m/\%d",
  date_angle = 0,
  popup_date = "\%a \%m/\%d/\%y \%I:\%M:\%OS3 \%p",
  bg = "white",
  fg = "black",
  displayModeBar = TRUE,
  height = 450,
  font.size = 12,
  shape.opacity = 0.4,
  shape.linecolor = "lightgray",
  line.width = 1,
  source = "source",
  register = TRUE,
  showticklabelsY = TRUE,
  config = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{data}{a timeline format dataframe containing named columns for
group, state, xleft, xright, ybottom, ytop. x coordinates
are datetime values and y coordinates are numeric indices indicating
vertical height of state lanes}

\item{auto.y}{(logical) whether to automatically compute lane
y coordinates or use ybottom/ytop arguments.}

\item{xrange}{optional, vector of date/times used to specify the
xrange of the plot. If NULL, the xrange defaults to the range inherent in
the min and max time values in the data: c(min(xleft), max(xright)). This
option is used to visualize the contexts within a broader time period.}

\item{collapse}{(logical) whether to draw grouped states in a single lane or
draw individual state activity in separate lanes.  Collapsing will result
in overlap of state rectangles unless states in the group are
mutually exclusive}

\item{yspace}{vertical space with value \link{0,1} to add between lanes}

\item{legend}{whether to draw a legend for the 0/1 states}

\item{date_breaks}{a string giving the datetime breaks for the plot
(e.g., "10 mins", "3 hours")}

\item{date_labels}{the format for the datetime axis labels}

\item{date_angle}{angle to position date axis labels}

\item{popup_date}{format for dates in the popups (tooltips)}

\item{bg}{the background color for the plot}

\item{fg}{the foreground color for the plot (text and axis lines)}

\item{displayModeBar}{If using plotly, whether to display a toolbar
over the plot or not}

\item{height}{a number giving the height of the plotly plot in pixels}

\item{shape.linecolor}{color of line around rectangles}

\item{line.width}{width of line around rectangles}

\item{source}{string identification for plot events}

\item{config}{logical, whether to configure the plot plot toolbar (may want
to skip this when configuring subplots)}
}
\description{
Plot an interactive timeline of state activity using plotly
}
\details{
This function creates a state timeline plot using the plotly library
the plotly \code{add_trace} function.   See \code{plot_state_timeline_ggplot} for
a version that uses the ggplot2 library, but also has an optio for converting
from a ggplot object to a plotly object using \code{ggplotly}.
}
\author{
Neil Klepeis
}
