% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_streams_calendar.R
\name{plot_streams_calendar}
\alias{plot_streams_calendar}
\title{Plot sensor data streams faceted as a calendar}
\usage{
plot_streams_calendar(
  data,
  area = TRUE,
  line = FALSE,
  x.lab = "Time",
  y.lab = "Value",
  title = "Sensor Stream Data",
  scales = "free_x"
)
}
\arguments{
\item{data}{Long-format sensor data frame for multiple streams
containing a 'Time' column, index columns, and Response and Value columns}

\item{by}{time unit used to create an index variable to use in faceting
data by time}
}
\value{
a ggplot object
}
\description{
Plot sensor data streams faceted as a calendar
}
\details{
Plot streams faceted in the shape of a calendar
Aggregates to hours.
}
\author{
Neil Klepeis
}
