% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_streams_ggplot.R
\name{plot_streams_ggplot}
\alias{plot_streams_ggplot}
\alias{plot_streams}
\title{Plot sensor data streams faceted by time using ggplot}
\usage{
plot_streams_ggplot(
  data,
  by = "1 day",
  by.format = "\%a, \%m/\%d",
  calendar = TRUE,
  facet.response = FALSE,
  area = TRUE,
  line = FALSE,
  step = FALSE,
  point = FALSE,
  facet = TRUE,
  pad = TRUE,
  numcols = 7,
  alpha = 0.8,
  xlimits = NULL,
  upper.limit = NA,
  labels = NULL,
  date_labels = "\%I\%p",
  breaks = NULL,
  date_breaks = "12 hour",
  show.legend = TRUE,
  scales = "free_x",
  x.lab = "Time",
  y.lab = "Value",
  title = paste("Sensor Streams Grouped by", by),
  expand.x = c(0, 0),
  expand.y = c(0, 0),
  mytheme = theme_streams_light,
  ...
)
}
\arguments{
\item{data}{Long-format sensor data frame for multiple streams
containing a 'Time' column, index columns, and Response and Value columns}

\item{by}{time unit used to create an index variable to use in faceting
data by time}

\item{by.format}{time format for strip lables}

\item{calendar}{logical, whether to create a calendar-style facet
by adding padding on either side to create dummy (blank) panels}

\item{area}{logical, whether to fill in area under the curve when plotting data}

\item{line}{logical, whether to draw lines when plotting data}

\item{step}{logical, whether to draw steps to represent data points}

\item{point}{logical, whethe to draw points when plotting data}

\item{facet}{logical, whether to create wrapped facets (panels) conditioned on the
'by' parameter}

\item{pad}{logical, whether to pad time series to nearest start or end of "by"
time period}

\item{numcols}{the number of columns to use when faceting}

\item{alpha}{the alpha value to draw data series (fills)}

\item{xlimits}{a 2-element vector of limits or a function taking defaults limits and outputting new limits
for the datetime x axis.}

\item{date_labels}{the tick labels to use for time on the x-axis}

\item{date_breaks}{the tick breaks to use for time on the x-axis}

\item{show.legend}{logical, whether to show a legend for area/line data series (Response)}

\item{scales}{the 'scales' argument to ggplot facets determining whether x and y axis sacles
are allowed to be "free" or remain "fixed" over all panels}

\item{x.lab}{label to use for the x-axis (Time)}

\item{y.lab}{label to use for the y-axis (Value)}

\item{title}{the title to use for the plot}
}
\value{
a ggplot object
}
\description{
Plot sensor data streams faceted by time using ggplot
}
\details{
This function is intended to be used in create attractive and easy to understand
time series plots faceted by a specified time grouping, defaulting to a single day.

If variables not Time, Response, or Value are in the passed data, the function
concatenates all columns not Time, Response or Value to a single index
variable that is used to facet data series in addition
to faceting by the specified time grouping.
}
\author{
Neil Klepeis
}
