% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_streams_hourly_aggregate.R
\name{plot_streams_hourly_aggregate}
\alias{plot_streams_hourly_aggregate}
\title{Plot column chart showing hourly averages across all days}
\usage{
plot_streams_hourly_aggregate(
  data,
  x.lab = "Hour",
  y.lab = "Value",
  fill.color = "blue",
  title = "Sensor Stream Hourly Means",
  breaks = c(0, 6, 12, 18),
  labels = c("Midnight", "6a", "Noon", "6pm")
)
}
\arguments{
\item{data}{Long format data frame containing a 'Time', a 'Response' variable
indicating the sensor stream and a 'Value' variable containing the stream
data values.}
}
\value{
a ggplot object
}
\description{
Plot column chart showing hourly averages across all days
}
\details{
This function produces a column chart showing the mean
level of the data stream by hour of the day, aggregated across
all available days.
}
\author{
Neil Klepeis
}
