% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_streams_peaks.R
\name{plot_streams_peaks}
\alias{plot_streams_peaks}
\title{Plot identified peaks in a time series}
\usage{
plot_streams_peaks(
  streams,
  peaks,
  mergedpeaks = NULL,
  baseline = NULL,
  threshold = NULL,
  merge.threshold = NULL
)
}
\arguments{
\item{streams}{a tibble of sensor data streams in standard long format
with Time, Response, and Value variables}

\item{peaks}{}
}
\value{
a ggplot object
}
\description{
This function plots peaks that have bee
identified in a time series
}
\details{
Create a ggplot of the streams superimposed with
identified peaks, threshold, and baseline
}
\author{
David Torres-Mendoza and Neil Klepeis
}
