% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeline_episodes_plotly.R
\name{plot_timeline_episodes_plotly}
\alias{plot_timeline_episodes_plotly}
\title{Create stacked plots of grouped data streams}
\usage{
plot_timeline_episodes_plotly(
  timeline,
  breaks,
  ncols = 3,
  height = NULL,
  bg = "white",
  fg = "black",
  ...,
  displayModeBar = TRUE
)
}
\arguments{
\item{timeline}{a dataframe containing group, state, xleft, xright}

\item{breaks}{a dataframe with start/end columns defining episodes}

\item{bg}{background color}

\item{fg}{foreground color}

\item{displayModeBar}{logical, whether to show the mode bar for the plot}
}
\description{
This function creates a stacked plotly combination plot
with individual subplots for each unique set of index values
}
\details{
The subplots are stacked on top of one another
}
\author{
Neil Klepeis
}
