% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantize_contexts.R
\name{quantize_contexts}
\alias{quantize_contexts}
\title{Quantize the times of a context sequence}
\usage{
quantize_contexts(..., unit = "10 seconds", method = "round")
}
\arguments{
\item{...}{one or more sequences of contexts to quantize}

\item{unit}{time interval to define the quantization}

\item{method}{one of "floor", "ceiling", or "round"}
}
\value{
Returns a list of all passed context sequences with
quantized times
}
\description{
Round times of a context sequence to tne nearest
grid value
}
\details{
"round" quantizes the context times to nearest time unit value
"floor" quantizes the context times DOWN to the nearest unit boundary
"ceiling" quantizes the context times UP to the nearest unit boundary
}
\seealso{
lubridate::round_date
}
\author{
Neil Klepeis
}
