% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_context.R
\name{reformat.context}
\alias{reformat.context}
\title{Reformat Context}
\usage{
reformat.context(time, states, allStates = NULL, format = "BinaryWide")
}
\arguments{
\item{time}{starting time for the new context (required)}

\item{states}{a list with group elements containing
vectors of active states for the new context occurring at 'time'}

\item{allStates}{optional list with groups elements containing
vectors of ALL states for use with Binary formats}

\item{format}{a string giving the format for the conversion.  See Details.}
}
\value{
a list or tibble containing the reformatted context data
}
\description{
This function converts a single context, consisting of a time and
list of grouped active states, to one of three different data frame formats:
"ActiveStates", "BinaryWide", or "BinaryLong"
}
\details{
The 'format' can be one of the following

 * ActiveStates - a data frame with a Time column and a column with concatenated strings showing active states:  group1:state1a,state1b group2:state2a,state2b, ...
 * BinaryWide - a wide data frame with a Time column and columns for each grouped state with 0 or 1 cell values (inactive or active)
 * BinaryLong - a long data frame with Time, Group, State, and Value columns where value is 0 or 1 (inactive or active)
}
\author{
Neil Klepeis
}
