% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_context.R
\name{sample_context}
\alias{sample_context}
\title{Sample Context Sequence}
\usage{
sample_context(context, times)
}
\arguments{
\item{context}{a context sequences with a Time variable}

\item{times}{times used to sample contexts}
}
\value{
Returns a context sequence ampled at the specified
times
}
\description{
Return context instances at arbitrary
specified times in a context sequence
}
\details{
A context sequence consists of state transitions occurring
at specific increasing times; each context's state configuration
persists until a new context is specified. A context persists until
a new context occurs with possible changes in state.

This function takes a set of arbitrary precise times and returns
the corresponding contexts that occur at these these.  Times
typically do not occur at the transitions but in the "in between"
times when a given context is active.

Sampled contexts in the sequence will contain the states specified
in the immediately preceding context transition.

This function operates on context sequences of any format,
i.e., Binary Long, Binary Wide, Active State, or Grouped state.
A Time variable must be present.

If any of the specified times are less than the minimum time
in the sequence, they are silently ignored. If times are greater
than the maximum time in the sequence, then the last context
in the sequence is repeatedly assigned for each time, i.e., we
assume that the last context transition specified occurs for
all foreseeable time.

Note:  Times do not have to be POSIXt objects (datetime objects)
but can represent any numeric time quantity (e.g.,
number of seconds, minutes, milliseconds, etc.).
}
\author{
Neil Klepeis
}
