% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_contexts.R
\name{sample_contexts}
\alias{sample_contexts}
\title{Sample Context Sequence}
\usage{
sample_contexts(contexts, times)
}
\arguments{
\item{contexts}{a context sequence with a Time variable}

\item{times}{times used to sample contexts}
}
\value{
Returns a new context sequence having the specified
times
}
\description{
Return contexts for arbitrary times from a
context sequence
}
\details{
A context sequence consists of contexts junction occurring
at specific increasing times. A context persists until
a new context occurs with changes in states.
This function takes a set of arbitrary times and returns
the corresponding contexts that occur at those times.

This function can operate on any context sequence format with a Time
variable, i.e., Binary Long, Binary Wide, Active State,
or Grouped state.

If any of the specific times are less than the minimum time
in the sequence, it is silently ignored.  If times are greater than
the maximum sequence time, then the last specific context
is assigned.
}
\author{
Neil Klepeis
}
