% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmented_elapsed_time.R
\name{segmented_elapsed_time}
\alias{segmented_elapsed_time}
\title{Computed a segmented elapsed time variable}
\usage{
segmented_elapsed_time(x, interval = 1, units = "min", group.var)
}
\arguments{
\item{x}{a data frame with a Time variable}

\item{interval}{the time interval in 'units' to use in computing time factors}

\item{units}{a string giving the units to use, e.g., "hour"s, mins", "secs",
in computing elapsed time and time factors}

\item{group.var}{a character vector of variables names to use in grouping}
}
\value{
a data frame with a new elapsed time variable and a time factor variable
}
\description{
Compute elapsed times and time factors with the "0" restarting at the
beginning of time segments defined by grouping variables
}
\details{
This function creates a elapsed time variable that resets to "0"
at the beginning of time segments defined by a set of specified
grouping variables.   The elapsed time has units defined by 'units'.
The time factor variable (computed using the 'time_factor' function) is assigned
over successive intervals defined by 'interval' and 'units'
}
\author{
Neil Klepeis
}
