% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_text_file.R
\name{split_text_file}
\alias{split_text_file}
\title{Split a text file on matched character string}
\usage{
split_text_file(
  file,
  split_pattern,
  comment_pattern = NULL,
  comment_character = "#"
)
}
\arguments{
\item{split_pattern}{pattern of characters to use in splitting the file}

\item{comment_pattern}{pattern to replace with a comment character, e.g., "#"}

\item{comment_char}{defaults to "#"}
}
\value{
Returns a list of character strings resulting from splitting the original
texte file
}
\description{
Read in a text file and split it into pieces based on
a specified string of characters
}
\details{
The list of character string can be read into a tibble using read_csv
This function is useful for pre-processing a text file of sensor data
that contains multiple experiments or runs that are separated by a fixed character string (tag)
}
\author{
Neil Klepeis
}
