% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t2m.R
\name{t2m}
\alias{t2m}
\title{convert hhmm format to elapsed minutes}
\usage{
t2m(x, increasing = TRUE, split = TRUE, zero = FALSE)
}
\description{
Convert time in hhmm format to minutes after midnight
}
\details{
Simple R function to convert time
in \verb{hhmm' 24-h format to minutes after midnight; if }increasing=TRUE', then the
times are interpreted as constantly
increasing (perhaps over more than
a single day) and the elapsed minutes
are calculated accordingly (i.e., by adding
i*1440 to times that occur on the i=0th,1st,2nd...
day).
}
\author{
Neil Klepeis
}
