% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ta_addtimes.R
\name{ta_addtimes}
\alias{ta_addtimes}
\title{Insert events at time points in a tact object}
\usage{
ta_addtimes(newtimes, breaks, events, add.na.events = FALSE)
}
\description{
Insert events at time points in a tact object
}
\details{
Function to add (insert) times to a time-activity object, adjusting the
event and breaks specifications accordingly.
The new times must be within the original time boundaries.  All
event sets are adjusted.
}
\author{
Neil Klepeis
}
