% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ta_coll.R
\name{ta_coll}
\alias{ta_coll}
\title{Collapse time-activity objet}
\usage{
ta_coll(breaks, events)
}
\description{
Collapse time-activity objet
}
\details{
Collapse a time activity object so that there are no
repeated event flags in a row.  There can be single
flags repeated at different places in the series as
long as another, different flag separates them.

In other words, we remove unnecessary time breaks in the
time activity object.  Think of it as removing the
unneeded "clutter".
}
\author{
Neil Klepeis
}
