% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ta_samp.R
\name{ta_samp}
\alias{ta_samp}
\title{Samps a portion of a tact object}
\usage{
ta_samp(t, breaks, events, set = 1)
}
\description{
Samps a portion of a tact object
}
\details{
Function to return event flags for arbitrary values of
continuous time input (i.e., sample an event series),
given the boundaries of a time-activity event time series and
the corresponding event flags (i.e., the event codes or labels) --
where event boundaries and flags take the form of a histogram
with "breaks" = event boundaries, and "counts" = event flags.
'breaks' can be a tact object (or coercable) with components
'breaks' and 'events'.  \verb{events' can be missing and will be filled in by }tact' command.
}
\author{
Neil Klepeis
}
