% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ta_subset.R
\name{ta_subset}
\alias{ta_subset}
\title{Extracts subset of a tact object}
\usage{
ta_subset(range, breaks, events, verbose = TRUE)
}
\description{
Extracts subset of a tact object
}
\details{
Function to return a time subset of a time-activity object.
Returns a tact object that is a portion of the passed object given
two-values for the subset time limits (`range')
Passed event boundaries and flags take the form of a histogram
with "breaks" = event boundaries, and "counts" = event flags.
'breaks' can also be a tact object (or coercable) with components
'breaks' and 'events'.
}
\author{
Neil Klepeis
}
