% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tact.R
\name{tact}
\alias{tact}
\title{Create a time-activity object (tact)}
\usage{
tact(breaks, events, hhmm = FALSE)
}
\description{
Create a tact object
}
\details{
Creates a tact (time-activity) object from a series
of numeric time breaks and one or more equal-length sets of event flags,
there should be one more break than events, and the times must be
strictly increasing.  The event flags can either be numeric
codes (possibly with names) or character labels.  If `hhmm=TRUE', then
the time breaks are interpreted as 24-h clocktime in 'HHMM' character format
where HH=hour and MM=minute (eg,14:12); otherwise the breaks are in
units of continuous time (e.g., either seconds, minutes, or hours).
The HHMM times are assumed to be increasing, and this cannot be
checked because, for example, "0012" could occur \emph{after} "1203"
on a following day.
}
\author{
Neil Klepeis
}
