% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tact_collapse.R
\name{tact_collapse}
\alias{tact_collapse}
\title{Collapse time-activity objet}
\usage{
tact_collapse(...)
}
\description{
Collapse time-activity objet
}
\details{
Collapse events in one or more time-activity objects or
a multiple time-activity object.  Adjacent identical
events are converted into a single event, with new time breaks
corresponding to the previous lowest and highest values.
}
\author{
Neil Klepeis
}
