% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tact_merge.R
\name{tact_merge}
\alias{tact_merge}
\title{merge two or more tact objects}
\usage{
tact_merge(..., intersection.only = TRUE)
}
\description{
merge two or more tact objects
}
\details{
Function to 'merge' a number of time-activity objects, or a
multiple time-activity object, into a single time-activity object
that contains the separate events sets in the passed
object(s).  No information is lost; new time breaks
are introduced where appropriate.  The returned \verb{tact' object has the time range of the largest range that all of the passed objects have in common. All }tact' objects, taken together,
must have some time in common.
}
\author{
Neil Klepeis
}
