% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_contexts_clean.R
\name{theme_contexts_clean}
\alias{theme_contexts_clean}
\title{A minimal blue ggplot theme for plotting contexts}
\usage{
theme_contexts_clean(
  panel.spacing = 0,
  tick.length = 0.2,
  axis.text.size = 11,
  axis.title.size = 32,
  axis.font = "sans"
)
}
\description{
A minimal blue ggplot theme for plotting contexts
}
\details{
Based on the default 'theme_hc' in the ggthemes package
with the 'theme_economist' scale colors.
}
\seealso{
scales_fill_contexts
scales_color_contexts
}
\author{
Neil Klepeis
}
