% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_contexts_dark.R
\name{theme_contexts_dark}
\alias{theme_contexts_dark}
\title{A minimal dark ggplot theme for plotting contexts}
\usage{
theme_contexts_dark(
  panel.spacing = 0,
  tick.length = 0.2,
  grid.color = NA,
  bg.plot = "black",
  axis.color.ticks = "white",
  axis.color = "white",
  axis.color.line = NA,
  strip.line.size = 1.3,
  bg.strip = "gray20",
  panel.line.size = 1.3,
  bg.panel = "gray10",
  border.panel.color = "black",
  border.strip.color = "black"
)
}
\description{
A minimal dark ggplot theme for plotting contexts
}
\details{
Based on the default 'theme_dark' in the ggthemes package
}
\seealso{
scales_fill_contexts
scales_color_contexts
}
\author{
Neil Klepeis
}
