% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_streams_clean.R
\name{theme_streams_clean}
\alias{theme_streams_clean}
\title{A minimal blue ggplot theme for plotting sensor streams}
\usage{
theme_streams_clean(
  panel.spacing = 0,
  tick.length = 0.2,
  axis.text.size = 11,
  axis.title.size = 32,
  axis.font = "sans"
)
}
\description{
A minimal blue ggplot theme for plotting sensor streams
}
\details{
Based on the default 'theme_hc' in the ggthemes package
with the 'theme_economist' scale colors.
}
\author{
Neil Klepeis
}
