% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_streams_dark.R
\name{theme_streams_dark}
\alias{theme_streams_dark}
\title{A dark ggplot theme for plotting sensor streams}
\usage{
theme_streams_dark(
  panel.spacing = 0,
  tick.length = 0.2,
  grid.color = NA,
  bg.plot = "black",
  axis.color.ticks = "white",
  axis.color = "white",
  axis.color.line = NA,
  strip.line.size = 1.3,
  bg.strip = "gray20",
  panel.line.size = 1.3,
  bg.panel = "gray20",
  border.panel.color = "black",
  border.strip.color = "black"
)
}
\description{
A dark ggplot theme for plotting sensor streams
}
\details{
Based on the default 'theme_dark' ggplot theme.
}
\author{
Neil Klepeis
}
