% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_streams_light.R
\name{theme_streams_light}
\alias{theme_streams_light}
\title{A light ggplot theme for plotting sensor streams}
\usage{
theme_streams_light(
  panel.spacing = 0,
  tick.length = 0.2,
  bg.panel = "gray80",
  bg.strip = "white",
  grid.color = NA,
  axis.color.ticks = "gray60",
  axis.color.line = NA,
  strip.line.size = 1.3,
  panel.line.size = 1.3,
  border.panel.color = "white"
)
}
\description{
A light ggplot theme for plotting sensor streams
}
\details{
Based on the default 'theme_gray' ggplot theme.
}
\author{
Neil Klepeis
}
