% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline2activestate.R
\name{timeline2activestate}
\alias{timeline2activestate}
\title{Convert contexts (episodes) in non-overlapping
timeline format to active-state context format}
\usage{
timeline2activestate(
  data,
  time1.var = "StartTime",
  time2.var = "EndTime",
  group.var = "Group",
  state.var = "State"
)
}
\arguments{
\item{data}{a date frame containing timelines of individual events/episodes}

\item{time1.var}{the expected name for the starting time column in 'data' (defaults to "TimeStart")}

\item{time2.var}{the expected name for the ending time column in 'data' (defaults to "TimeEnd")}

\item{group.var}{the expected name for the group column in 'data' (defaults to "Group")}

\item{state.var}{the expected name for the state column in 'data' (defaults to "State")}
}
\description{
Convert contexts (episodes) in non-overlapping
timeline format to active-state context format
}
\details{
This function converts a timeline-formatted dataframe with Start
and End time values, and Group and State columns, to an active state
format context data frame with a Time column and columns containing
the Group:State value

IMPORTANT:  This function cannot handle overlapping episodes.
All episodes must be non-overlapping in time (they \emph{can} have overlapping endpoints).
This restriction is not currently checked.
}
\author{
Neil Klepeis
}
