% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xytact.R
\name{xytact}
\alias{xytact}
\title{Plot sensor straams and contexts based on lattice \code{xyplot} function}
\usage{
xytact(
  x,
  data,
  groups = NULL,
  tact.var,
  tact.data = data,
  codes,
  codes.colors,
  color.fun = gray.colors,
  group.factor = 1,
  key.space = "right",
  key.columns = 1,
  alpha.tact = 0.2,
  col = NULL,
  fill = TRUE,
  pattern = FALSE,
  pattern.type = c("segments", "points"),
  pattern.angle = 45,
  pattern.width = 3,
  pattern.col = "gray20",
  debug = FALSE,
  abline = NULL,
  show.outside.key = TRUE,
  show.inside.key = TRUE,
  show.labels = FALSE,
  ...
)
}
\arguments{
\item{x}{formula indicating X and Y variables and conditioning variables for the sensor streams}

\item{data}{a data frame containing the variables included in the formula}

\item{tact.var}{a character vector with names of variables in \code{data} or \code{tact.data}
that represent the time-activity of states with one group of mutually exclusive states
per variable}

\item{tact.data}{a data set (if other than \code{data}) containing time-activity data}
}
\description{
This function creates a lattice plot with continuous sensor streams and
discrete context overlays (timelines)
}
\details{
This is a legacy function to plot sensor streams along with contexts in the form
of rectangles overlaid onto lattice panels to represent timelines of states.
This kind of plot may be possible using ggplot facets but there seems to
be more flexible to create conditioning panel with overlays using the
\code{lattice} R package.
}
\author{
Neil Klepeis
}
