% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_context.R
\name{as_context}
\alias{as_context}
\title{Coerce data frame or tibble to a long-format 'context' data object}
\usage{
as_context(x)
}
\arguments{
\item{data}{a dataframe or tibble containing Time, State, Value, and, optionally
other index variables to convert to a 'context' data type}
}
\description{
Coerce data frame or tibble to a long-format 'context' data object
}
\details{
Returns a 'context' data object, a tibble with specific cols types, or
an error if the passed object cannot be coerced to a 'context' data type
}
\author{
Neil Klepeis
}
