% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin2lnorm.R
\name{bin2lnorm}
\alias{bin2lnorm}
\title{Fit log-normal probability model to binned data (histogram)}
\usage{
bin2lnorm(
  l,
  h,
  wt,
  plot = FALSE,
  density = FALSE,
  mcol = "red",
  dcol = "black",
  lty = "solid",
  lwd = 3,
  shade.density = 15,
  shade.angle = 35,
  rect.border.col = NULL,
  rect.lwd = 1,
  rect.lty = 1,
  cex = 1.4,
  main,
  xlab,
  ylab,
  ...
)
}
\arguments{
\item{l}{class interval limits}

\item{h}{"height" values of the histogram (counts)}

\item{density}{logical, whether to plot the density function}
}
\value{
A list containing
\code{h}, \code{l}, \code{log(l)},
the probability density (PDF), the cumulative probabilities (CDF),
the normal quartiles, the geometric mean (GM), and the
geometric standard deviation (GSD)
}
\description{
This function uses weighted least squares to fit a log-normal
probability model to binned data, i.e., counts of data points within
specific class intervals (a histogram)
}
\details{
As a side effect the function will plot the
result of the fit either as a density plot (probability versus value)
or as a cumulative distribution function (CDF).

This function takes binned data (i.e., a histogram) and
fits a lognormal model to it, using weighted least-squares,
and optionally plotting the fit and the data together as a CDF.
If density = TRUE then it will plot a histogram with
the fitted lognormal function superimposed, otherwise it plots
a cdf (log-probability plot) with the fitted line through the
data points.

This is similar to the 'method of quantiles' for
estimating the parameters of a log-normal distributtion,
except that we use all of the data excluding the top point. See
Ott, 1995, page 268, "Environmental Statistics and Data
Analysis".
}
\author{
Neil Klepeis
}
