% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_contexts.R
\name{combine_contexts}
\alias{combine_contexts}
\title{Combine two or more context sequences}
\usage{
combine_contexts(..., replace = FALSE)
}
\arguments{
\item{...}{context sequence data frames in BINARY WIDE format}

\item{replace}{whether to merge with prior sequenes or replace sequences}
}
\value{
a data frame containing a sequence consisting of merged or
replaced conetxt sequences
}
\description{
Combine two or more context sequences into a
single context sequences with following contexts either merging
with prior sequences or replacing sequences
}
\details{
Iterating from i = 2 to n sequences we modify the i=1 sequence to
include all times and sequences from the following sequences

This consists of standardizing context groups/states, standardizing
the times, and then adding context values for aligned rows of
different context sequences

After we have the sames groups/states in all sequences (inserting 0's for missing ones)
Iterate from i = 2 to n:
(1) Add times of ith sequence to i = 1 sequence
(2) Add times of i = 1 sequence to ith sequence
(3) Sum the context values for aligned rows

Could we just use a left_join or full_join to
combine all of these together, if there
are duplicated, then we sum them togehter. maybe..no

For replacements, perhaps we remove all the contexts first
before doing the join... between min/max time of following
context sequences
}
\examples{

library(tidyverse)
}
\author{
Neil Klepeis
}
