% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_binary_contexts.R
\name{convert_binary_contexts}
\alias{convert_binary_contexts}
\title{Convert Binary Context}
\usage{
convert_binary_contexts(bc, fromLong = TRUE, toBinary = TRUE, sep = ":")
}
\arguments{
\item{bc}{a data frame containing binary contexts (the default storage format is LONG)}

\item{toBinary}{whether format of returned value is binary or active-state format}

\item{from}{format of 'bc' argument, "long" or "wide" binary format}
}
\description{
This function converts a context in binary format
to an alternative binary format (long or wide) or to active-state
format (timestamped list of active states)
}
\details{
When converting to active-state format, state values >=1 are
recognized as part of "active" contexts
}
\author{
Neil Klepeis
}
