% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_from_activestate.R
\name{convert_from_activestate}
\alias{convert_from_activestate}
\title{Convert Contexts from Active State Format}
\usage{
convert_from_activestate(contexts, allStates, sep = "|")
}
\arguments{
\item{contexts}{tibble containing contexts in active-state format}

\item{allStates}{list containing group elements with all states to match}

\item{sep}{the separator between contexts, defaults to "|"}
}
\value{
a tibble containing contexts in binary long format
}
\description{
This function converts a data frame of contexts in active-state
format to binary long format
}
\details{
The input active-state contexts should have a Time
column, a States column, and any additional index columns.
}
\author{
Neil Klepeis
}
