% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_time_segment_from_contexts.R
\name{cut_time_segment_from_contexts}
\alias{cut_time_segment_from_contexts}
\title{Cut a time segment from contexts}
\usage{
cut_time_segment_from_contexts(contexts, start, stop)
}
\arguments{
\item{contexts}{a sequence of any number of contexts}

\item{start}{beginning time to start cutting}

\item{end}{end time to stop cutting}
}
\value{
Returns a new context sequence with a context segment
removed
}
\description{
Removes all contexts in a context sequence between
the specified times
}
\author{
Neil Klepeis
}
