% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_stream_episodes.R
\name{find_stream_episodes}
\alias{find_stream_episodes}
\title{Find Stream Episodes}
\usage{
find_stream_episodes(
  streams,
  threshold,
  context = TRUE,
  group = "Episodes",
  state = "Episode"
)
}
\arguments{
\item{streams}{a tibble of sensor data streams in standard
Time, Response, Value long format}

\item{threshold}{the response value to use as a threshold when
identifying peaks and the extent of each episode}

\item{context}{if TRUE return in active-state context format
other return a tibble with start/end times for each episode}
}
\value{
a vector of breaks identifying stream episodes
}
\description{
This function identifies episodes in a
stream of sensor data
}
\details{
A stream episode is defined as a segment of time
in which data stream values contiuously occur above a specified
threshold. If multiple responses are present in the passed
value only the first one is used to identify episodes
}
\author{
Neil Klepeis
}
