% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_log_normal.R
\name{fit_log_normal}
\alias{fit_log_normal}
\title{Fit a log normal distribution to data}
\usage{
fit_log_normal(y, forceNA = TRUE, plot = TRUE)
}
\arguments{
\item{...}{}
}
\description{
Fit a log normal probability distribution to data
}
\author{
Neil Klepeis
}
