% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_log_probability_bins.R
\name{geom_log_probabilility_bins}
\alias{geom_log_probabilility_bins}
\alias{geom_log_probability_bins}
\title{Add binned data to log probability plot}
\usage{
geom_log_probability_bins(
  mapping = NULL,
  data = NULL,
  stat = "identify",
  position = "identity",
  ...,
  breaks,
  counts,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{counts}{}
}
\description{
Add binned data to an exisiting log-probability plot
}
\author{
Neil Klepeis
}
