% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_files.R
\name{list_data_files}
\alias{list_data_files}
\title{Get a list of data files for processing}
\usage{
list_data_files(
  prefixes = c("CNR", "IPS", "SidePak", "DustTrak"),
  middle = "",
  extension = "csv",
  full.names = TRUE,
  recursive = TRUE
)
}
\arguments{
\item{prefixes}{prefix(es) for matching files, defaults ot "CNR", "IPS", "PurpleAir", "SidePak", "DustTrak"}

\item{extension}{extension for matching files, defaults to "csv"}

\item{full.names}{logical, whether to return full names with paths, default to TRUE}

\item{recursive}{logical, whether to look into subdirectories recursively to match files, defaults to TRUE}
}
\value{
list of data file names matching prefix and extension
}
\description{
Look in current directory for files with specified prefixes
and extension
}
\details{
By default looks recursively in sub-directories and returns full names
with the file path relative to current directory
}
\author{
Neil Klepeis
}
