% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_historical_data_plotly2.R
\name{plot_historical_data_plotly2}
\alias{plot_historical_data_plotly2}
\title{Plot LONG historical sensor data using the plotly package}
\usage{
plot_historical_data_plotly2(
  data,
  area = TRUE,
  title = NULL,
  showlegend = TRUE,
  legend.inside = TRUE,
  shape = "hv",
  source = "source",
  fg = "white",
  bg = "blue",
  line.width = 1,
  register = TRUE,
  include.rangeSelector = FALSE,
  include.rangeSlider = FALSE
)
}
\arguments{
\item{data}{Long-format sensor data frame for multiple streams frame
containing a 'Time' column, index columns, and Response and Value columns}

\item{area}{whether to fill in the plot, defaults to TRUE}

\item{title}{title to include (NULL by default)}

\item{showlegend}{whether to include legend,defaults to TRUE}

\item{legend.inside}{whether to include legend inside the plot, defaults to TRUE}

\item{shape}{what shape to use with the line, defaults to "hv"}

\item{source}{what source to assign to the plot, defaults to "source"}

\item{fg}{foreground color, defaults to white}

\item{bg}{background color, defaults to blue}

\item{line.width}{plot line.width defaults to 1}

\item{register}{whether to register the plot for updates, defaults to TRUE}

\item{include.rangeSelector}{whether to include the range selector, defaults to FALSE}

\item{include.rangeSlider}{whether to include the range slider, defaults to FALSE}
}
\value{
a plotly object
}
\description{
Plot LONG historical sensor data using the plotly package
}
\details{
concatenates all columns not Time or Value to
a single index variable that is used to plot data series.

See 'plot_multiple_historical_dataL.R' which calls this function
to produce stacked subplots for each non-Response index variable, i.e.,
all index variables besides Response are used to facet the data into
separated plots
}
\author{
Neil Klepeis
}
