% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_log_probability.R
\name{plot_log_probability}
\alias{plot_log_probability}
\title{Log probability plot}
\usage{
plot_log_probability(xtics, ytics)
}
\arguments{
\item{xtics}{vector of x-axis (probability) tic values}
}
\description{
Create an empty log-probability plot using ggplot
}
\details{

}
\author{
Neil Klepeis
}
