% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_streams_peaks.R
\name{plot_streams_peaks}
\alias{plot_streams_peaks}
\title{Plot identified peaks in a time series}
\usage{
plot_streams_peaks(x)
}
\arguments{
\item{x}{list containing the output of the find_streams_episodes
function with values for streams, peaks, merged peaks,
and parameters}
}
\value{
a ggplot object
}
\description{
This function plots peaks that have bee
identified in a time series
}
\details{
Create a ggplot of the streams superimposed with
identified peaks, threshold, and baseline
}
\author{
David Torres-Mendoza and Neil Klepeis
}
