% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantize_contexts.R
\name{quantize_contexts}
\alias{quantize_contexts}
\title{Quantize the times of a context sequence}
\usage{
quantize_contexts(contexts, grid)
}
\arguments{
\item{contexts}{a sequence of any number of contexts}

\item{grid}{time interval to define the quantization grid}
}
\value{
Returns a new context sequence with quantized times
}
\description{
Round times of a context sequence to tne nearest
grid value
}
\author{
Neil Klepeis
}
