% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_context.R
\name{sample_context}
\alias{sample_context}
\title{Sample Context Sequence}
\usage{
sample_context(context, times)
}
\arguments{
\item{context}{a context sequences with a Time variable}

\item{times}{times used to sample contexts}
}
\value{
Returns a context sequence ampled at the specified
times
}
\description{
Return context instances at arbitrary
specified times from a context sequence
}
\details{
A context sequence consists of state transitions occurring
at specific increasing times; each context's state configuration
persists until a new context is specified. A context persists until
a new context occurs with possible changes in state.

This function takes a set of arbitrary times and returns
the corresponding contexts that occur at precise times.

Sampled contexts in the sequence will duplicate the states in
the immediately preceding context.

This function can operate on any context sequence format with a Time
variable, i.e., Binary Long, Binary Wide, Active State,
or Grouped state.

If any of the specified times are less than the minimum time
in the sequence, they are silently ignored. If times are greater than
the maximum time in the sequence, then the last context
in the sequence is repeatedly assigned for each time.

Note:  Times do not have to be Time.POSIX objects but can
represent any numeric time quantity (e.g., number of seconds,
minutes, etc.).
}
\author{
Neil Klepeis
}
